refactor: api からデータをfetch してきている部分の関数の管理を変更
コンポーネント内から除外するため
